+++ include/aligator/modelling/control-box-function.hpp
@@ -3,18 +3,19 @@
 #include "aligator/core/function-abstract.hpp"
 
 namespace aligator {
-/**
- * @brief   A simple function \f$f(u) = [u_{\min} - u; u - u_{\max}]\f$.
- *
- * @details This function should be used along
- * proxsuite::nlp::NegativeOrthantTpl to create control bound constraints \f[
- * -u_\min \leq u \leq u_\max. \f]
- */
+///
+/// @brief   A simple function \f$f(u) = [u_{\min} - u; u - u_{\max}]\f$.
+///
+/// @note This function was initially meant to be used along with
+/// NegativeOrthantTpl to create control bound constraints.
+/// For this purpose, please use BoxConstraintTpl instead.
+/// @deprecated This class is not meant to be used anymore, and has been
+/// deprecated. It might be removed in the future. Use BoxContraintTpl to
+/// implement control bounds. See related note.
 template <typename _Scalar>
-struct ALIGATOR_DEPRECATED_MESSAGE(
-    "ControlBoxFunction should not be used. Instead, just use the identity "
-    "function and a BoxConstraint.") ControlBoxFunctionTpl
-    : StageFunctionTpl<_Scalar> {
+struct [[deprecated("ControlBoxFunction should not be used. Instead, just use "
+                    "the identity function and a BoxConstraintTpl.")]]
+ControlBoxFunctionTpl : StageFunctionTpl<_Scalar> {
   using Scalar = _Scalar;
   ALIGATOR_DYNAMIC_TYPEDEFS(Scalar);
   using Base = StageFunctionTpl<Scalar>;